# perf: 1.8x or greater speed improvements in proof recursion